swinir/requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
facenet-pytorch==2.6.0
filelock==3.17.0
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.0
scipy==1.15.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tifffile==2025.1.10
timm==1.0.14
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
